refactor(layout): reorganize layout components into a separate folder called layout
