+++ src/App.tsx
@@ -1,7 +1,19 @@
-import { Layout } from "@/components";
+import { Main } from "@/layout";
+import { AppProvider, PaletteProvider, ToolsProvider } from "@/context";
+import { BrowserRouter } from "react-router";
 
 function App() {
-    return <Layout />;
+    return (
+        <BrowserRouter>
+            <AppProvider>
+                <PaletteProvider>
+                    <ToolsProvider>
+                        <Main />
+                    </ToolsProvider>
+                </PaletteProvider>
+            </AppProvider>
+        </BrowserRouter>
+    );
 }
 
 export default App;
