+++ src/components/index.ts
@@ -1,7 +1,6 @@
 export * from "./colors";
 export * from "./input";
 export * from "./label";
-export * from "./layout";
 export * from "./picker";
 export * from "./preview";
 export * from "./popover";
@@ -16,6 +15,4 @@ export * from "./logo";
 export * from "./combo-box";
 export * from "./scroll";
 export * from "./zoom";
-export * from "./sidebar";
 export * from "./radio-group";
-export * from "./footer";
