+++ src/layout/footer/types.ts
