+++ src/layout/index.ts
@@ -1,0 +1,3 @@
+export * from "./main";
+export * from "./sidebar";
+export * from "./footer";
